refactor: move scenarios to policy, add placeholders with DeprecationWarning
- scenarios/ -> policy/scenarios/ (abstract_scenario, reform_scenario)
- Placeholders at openfisca_survey_manager/scenarios/ with DeprecationWarning
- Remove common, processing/weights, root placeholders (coicop, matching, etc.)
- Update all imports to policy.scenarios
- Add missing Survey import in abstract_scenario

Made-with: Cursor
