+++ openfisca_survey_manager/policy/scenarios/__init__.py
@@ -1,0 +1,4 @@
+from openfisca_survey_manager.policy.scenarios.abstract_scenario import AbstractSurveyScenario
+from openfisca_survey_manager.policy.scenarios.reform_scenario import ReformScenario
+
+__all__ = ["AbstractSurveyScenario", "ReformScenario"]
