+++ openfisca_survey_manager/processing/__init__.py
@@ -6,7 +6,8 @@ from openfisca_survey_manager.processing.harmonization import harmonize_data_fra
 # Lazy import to avoid circular dependency (processing -> policy -> survey_collections -> core)
 def __getattr__(name: str):
     if name in ("Calibration", "calmar", "check_calmar"):
-        from openfisca_survey_manager.processing.weights import Calibration, calmar, check_calmar
+        from openfisca_survey_manager.policy.calibration import Calibration
+        from openfisca_survey_manager.policy.calmar import calmar, check_calmar
 
         return {"Calibration": Calibration, "calmar": calmar, "check_calmar": check_calmar}[name]
     raise AttributeError(f"module {__name__!r} has no attribute {name!r}")
