+++ openfisca_survey_manager/tests/test_calibration.py
@@ -2,7 +2,7 @@ from openfisca_core import periods
 from openfisca_core.tools import assert_near
 
 from openfisca_survey_manager.policy.calibration import Calibration
-from openfisca_survey_manager.scenarios.abstract_scenario import AbstractSurveyScenario
+from openfisca_survey_manager.policy.scenarios.abstract_scenario import AbstractSurveyScenario
 from openfisca_survey_manager.tests import tax_benefit_system
 from openfisca_survey_manager.tests.test_scenario import (
     create_randomly_initialized_survey_scenario,
