+++ openfisca_survey_manager/tests/test_enum.py
@@ -1,7 +1,7 @@
 import pandas as pd
 from openfisca_country_template.variables.housing import HousingOccupancyStatus
 
-from openfisca_survey_manager.scenarios.abstract_scenario import AbstractSurveyScenario
+from openfisca_survey_manager.policy.scenarios.abstract_scenario import AbstractSurveyScenario
 from openfisca_survey_manager.tests import tax_benefit_system
